requirements.txt
fastapi==0.115.0
uvicorn==0.35.0
pandas==2.2.3
numpy==2.3.3
chromadb==1.0.12
langchain==0.3.27
langchain-core==0.3.76
sentence-transformers==5.1.0
huggingface-hub==0.34.4
python-dotenv==1.1.1
